refactor(PresentProductList) : ui 생성이라는 역할을 관리한다는 관점으로 props로 data를 전달받아 ui를 생성하도록 재정의
